fixed: coverage failing due to my improperty set up of the initial mathfield testing file
